TradePoint Type 수정, Reward 랜덤박스 신청 Test 완료 중간 커밋

+++ src/main/java/city/olooe/plogging_project/dto/RewardDTO.java
@@ -6,6 +6,8 @@ import java.util.Date;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 
+import org.apache.ibatis.javassist.compiler.ast.Member;
+
 import city.olooe.plogging_project.model.DonationEntity;
 import city.olooe.plogging_project.model.MemberEntity;
 import city.olooe.plogging_project.model.ProductEntity;
@@ -32,7 +34,7 @@ public class RewardDTO {
     // private LocalDateTime localDateTime;
     private Long rewardNo; // PK, Reward 번호
     private String type; // 포인트 유형(기부 or 랜덤박스)
-    private double tradePoint; // 포인트 증감, 차감액
+    private Long tradePoint; // 포인트 증감, 차감액
     private Date rewardDate; // 포인트 업데이트 시간
     private Long memberNo; // FK, 회원 번호
     private Long dno; // FK, 기부 번호
@@ -58,5 +60,4 @@ public class RewardDTO {
                 .productEntity(ProductEntity.builder().pno(dto.getPno()).build())
                 .build();
     }
-
 }
